Update userscript.js
Adjusted styles to remove the need for ul, and li elements by applying the responsive column properties directly to inputs, rather than wrapping them in list items. Still need to clean up the css & check to see if I still need list item CSS definitions. The goal right now is to remove them if they aren't required.

Also, created a filter ability for the function library (funcLib) to filter out which of it's methods to convert to buttons. The goal was to skip the toggle button for the textarea, since it is manually applied in the textarea module. What was the reason for putting the toggle button is in the funcLib in the first place? I could not seem to trigger the function due to issues with the scope. The funcLib is designed to dynamically map functions to buttons, so I needed to add a filter to keep the toggle button from showing up twice in the menu. For future modules, some way of hooking into the funcLib from another script would allow for extensions to be built. Right now, the idea is to create column-based modules, since the size can be adjusted to take up any number of the 12 column span.