Update userscript.js
Named the 3 text inputs: "a,b,c". Named 3 number inputs: "x,y,z". Using generic variables to make the usable by many modular functions. The next challenge is providing dynamic labels for the inputs. It would be extremely helpful to know how the inputs are used in each module. Modules are generally triggered by buttons. The buttons are created dynamically from the names of each function in the library (funcLib). The current plan is update the placeholders when a button is hovered or focused. On hover, map the variable names and default values of expected user inputs. Maybe use a conditional in the original function to check if the action is "read" or "execute".